requirements.txt
streamlit
tensorflow
numpy
pandas
matplotlib
pillow